// src/app/core/enums/control-type.enums.ts
export enum ControlTypeMode {
  Text = 1,
  Number = 2,
  Textarea = 3,
  SingleSelect = 4,
  MultipleSelect = 5,
  RadioButton = 6,
  Checkbox = 7,
  Date = 8,
  File = 9,
  Image = 10,
  UserSelect = 11,
  Email = 12,
  Phone = 13,
  DateRange = 14,
  newStep = 15,
  UserSelectMultiple = 16,
  RichText = 17,
  DownloadTemplate = 19,
  repeater = 20,
}